renamed stream_iteration to log_iteration and stream_timed_iteration to log_timed_iteration
